wip to add AWS
the steps here will do a full creation of the cluster, which include
VPC, subnets (private and public), and security group, associating (and
creating if needed) a pem, getting the endpoint and certificate to make
a kube config yaml file to authenticate, and then another stack to
create the workers pool. This is interesting that AWS first creates you
an "empty" cluster, meaning just a control plane, and then you need
to create the workers as a separate request, and apply a config map
secret to the kube-system so the control plane can see the workers!
This is so much more complex than GKE, and now that I have everything
working to go UP I have to go backwards and figure out how to delete
everything before looking into scaling... hahahahahahah aahhhhh! :)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>