fix waiting for nodes (#6)
* fix waiting for nodes

we currently rely on subprocess with eksctl, and this
is not ideal for non-local environments (e.g., inside
the flux operator). With this updated approach we
instead generate a token with an STS handle, and instan-
tiate an ApiClient configuration and pythonic kubectl.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* add note about checking for creation

and also add note about credential consistency. The user
should choose a long expiring environment set OR the
$HOME/.aws/config but ideally not both.

* add more verbose logging messages

Signed-off-by: vsoch <vsoch@users.noreply.github.com>