testing updated algorithm for scaling
We currently will go too big because the assumption is scaling
by just 1, so we need to update our scaling functions (now renamed
to increase_by and decrease_by to indicate returning an actual value)
to return the value to actually increment/change by, and then do it.
This is a WIP for testing.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>