wip to add better support for node groups/autoscaling on gke
This is still risky to do - the GKE clusters self update
and often render in a state that is not usable. But this is still better than
the current setup. When I am brave I will try bringing this up again.
Another issue here is by using a region, we set a min/max for EACH region
so it can result in 3-4x the number of nodes that are expected (not
a great surprise imho).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>